requirements.txt
# Core PDF Processing
pdfplumber==0.11.0
pdf2image==1.17.0
pytesseract==0.3.10
Pillow>=10.0.0,<11.0.0
click==8.1.7
python-dateutil==2.9.0.post0
reportlab==4.2.2
camelot-py==0.11.0
opencv-python>=4.8,<5.0
tabula-py==2.9.0

# Streamlit App Dependencies
streamlit>=1.32.0,<2.0.0
boto3>=1.34.0,<2.0.0
pandas>=2.2.0,<3.0.0
openpyxl>=3.1.2,<4.0.0
PyMuPDF>=1.26.4,<2.0.0
pymupdf4llm==0.0.17
openai>=1.30.0,<2.0.0
altair>=5.3.0,<6.0.0
tiktoken>=0.7.0,<1.0.0


